refactor groups and keys package
