+++ auth/api/http/doc.go
@@ -1,0 +1,3 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+package http
