+++ auth/api/http/keys/endpoint_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package http_test
+package keys_test
 
 import (
 	"context"
