+++ auth/api/http/keys/requests.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package http
+package keys
 
 import (
 	"time"
